requirements.txt
streamlit==0.62.0
pandas
plotly.express
sklearn
numpy
matplotlib
pickle

